code clean up

+++ core/vm/arch/jit/amd64/jit_amd_lp64.h
@@ -731,17 +731,17 @@ namespace Runtime {
     // move instructions
     void move_reg_mem8(Register src, long offset, Register dest);
     void move_mem8_reg(long offset, Register src, Register dest);
-    void move_imm_mem8(long imm, long offset, Register dest);
+    void move_imm_mem8(int8_t imm, long offset, Register dest);
     void move_reg_mem16(Register src, int32_t offset, Register dest);
     void move_mem16_reg(int32_t offset, Register src, Register dest);
-    void move_imm_mem16(int32_t imm, int32_t offset, Register dest);
+    void move_imm_mem16(int16_t imm, int32_t offset, Register dest);
     void move_imm_mem32(int32_t imm, int32_t offset, Register dest);
     void move_reg_mem32(Register src, long offset, Register dest);
     void move_mem32_reg(long offset, Register src, Register dest);
     void move_reg_reg(Register src, Register dest);
     void move_reg_mem(Register src, long offset, Register dest);
     void move_mem_reg(long offset, Register src, Register dest);
-    void move_mem_reg32(int32_t offset, Register src, Register dest);
+    void move_mem_reg32(long offset, Register src, Register dest);
     void move_imm_memx(RegInstr* instr, long offset, Register dest);
     void move_imm_mem(long imm, long offset, Register dest);
 #ifdef _WIN64
